+++ src/commands/wholeft.h
@@ -1,0 +1,8 @@
+#ifndef CMD_WHOLEFT_H
+#define CMD_WHOLEFT_H
+
+__SWIRC_BEGIN_DECLS
+void cmd_wholeft(CSTRING);
+__SWIRC_END_DECLS
+
+#endif